requirements.txt
GitPython
progress
anybadge
Pillow
imgrender
matplotlib
seaborn
matplotlib
seaborn
wordcloud
nltk
plotly
emoji
numpyencoder
urllib3
bs4
# heatmap
PyGithub
